Split out header renderer
